[#1512] Add onetime/constants, rebalance onetime
